requirements.txt
requests
bs4
secure-smtplib
python-dotenv